- Mike Gist fixed some compile warnings.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@27784 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ libs/csgeom/bsptree.cpp
@@ -189,7 +189,7 @@ void csBSPTree::Build (CS::TriangleIndicesStream<int>& triangles,
     triidx.Push (int (tris.Push (t)));
   }
 
-  Build (tris.GetArray(), planes.GetArray(), tris.GetSize(), vertices, triidx);
+  Build (tris.GetArray(), planes.GetArray(), (int)tris.GetSize(), vertices, triidx);
 }
 
 void csBSPTree::Build (csTriangle* triangles, int num_triangles,
